before breaking changes

+++ mzLib/MassSpectrometry/IRetentionTimeAlignable.cs
@@ -2,6 +2,6 @@
 public interface IRetentionTimeAlignable
 {
     public string FileName { get; set; }
-    public double RetentionTime { get; set; }
+    public float RetentionTime { get; set; }
     public string Identifier { get; }
 }
